requirements.txt
Django>=6.0.4
requests>=2.31.0
django-allauth>=0.61.0
python-dotenv>=1.0.0
Pillow>=10.0.0
pymongo[srv]>=4.6.0
dnspython>=2.4.0
psycopg>=3.1.18
psycopg-binary>=3.1.18
djangorestframework>=3.15.1
whitenoise>=6.6.0
django-extensions>=3.2.3
django-cors-headers>=4.3.0  